+++ core/hoyo/stamina.js
@@ -5,12 +5,6 @@ export default class Stamina extends HoyoTemplate {
 	static MAX_STAMINA = 180;
 	static DS_SALT = "6s25p5ox5y14umn1p61aqyyvbvvl3lrt";
 
-	static timeUnits = {
-		h: { m: 60, s: 3600, ms: 3600.0e3 },
-		m: { s: 60, ms: 60.0e3 },
-		s: { ms: 1.0e3 }
-	};
-
 	static data = new Map();
 
 	static async checkAndRun (options = {}) {
@@ -18,7 +12,7 @@ export default class Stamina extends HoyoTemplate {
 
 		for (const [uid, account] of Stamina.data) {
 			const generatedDS = Stamina.generateDS(Stamina.DS_SALT);
-			const region = Stamina.getAccountRegion(account.uid);
+			const region = sr.Utils.getAccountRegion(account.uid);
 
 			const res = await sr.Got({
 				url: "https://bbs-api-os.hoyolab.com/game_record/hkrpg/api/note",
@@ -67,7 +61,7 @@ export default class Stamina extends HoyoTemplate {
 				stamina_recover_time: staminaRecoverTime
 			} = data;
 
-			const delta = Stamina.formatTime(staminaRecoverTime);
+			const delta = sr.Utils.formatTime(staminaRecoverTime);
 			const objectData = {
 				uid,
 				currentStamina,
@@ -117,7 +111,7 @@ export default class Stamina extends HoyoTemplate {
 	static async loadData () {
 		const accounts = sr.Account.getActiveAccounts();
 		for (const account of accounts) {
-			if (!this.getAccountRegion(account.uid)) {
+			if (!sr.Utils.getAccountRegion(account.uid)) {
 				continue;
 			}
 
@@ -155,47 +149,5 @@ export default class Stamina extends HoyoTemplate {
 			.createHash("md5")
 			.update(string)
 			.digest("hex");
-	}
-
-	static formatTime (seconds = 0) {
-		const array = [];
-
-		if (seconds >= Stamina.timeUnits.h.s) {
-			const hours = Math.floor(seconds / Stamina.timeUnits.h.s);
-			seconds -= hours * Stamina.timeUnits.h.s;
-			array.push(`${hours} hr`);
-		}
-        
-		if (seconds >= Stamina.timeUnits.m.s) {
-			const minutes = Math.floor(seconds / Stamina.timeUnits.m.s);
-			seconds -= minutes * Stamina.timeUnits.m.s;
-			array.push(`${minutes} min`);
-		}
-
-		if (seconds >= 0 || array.length === 0) {
-			array.push(`${this.round(seconds, 3)} sec`);
-		}
-
-		return array.join(", ");
-	}
-
-	static round (number, precision = 0) {
-		return Math.round(number * (10 ** precision)) / (10 ** precision);
-	}
-
-	static getAccountRegion (uid) {
-		const region = uid.toString().slice(0, 1);
-		switch (region) {
-			case "9":
-				return `prod_official_cht`;
-			case "8":
-				return `prod_official_asia`;
-			case "7":
-				return `prod_official_eur`;
-			case "6":
-				return `prod_official_usa`;
-			default:
-				return false;
-		}
 	}
 }